requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
google-generativeai==0.3.2
python-multipart==0.0.9
python-dotenv==1.0.1
langchain==0.0.267
openai==0.27.8
httpx==0.24.1
aiohttp==3.9.3
requests==2.31.0
websockets==11.0.3
pytest==7.4.2
pytest-asyncio==0.21.1
asyncio==3.4.3
# Added for multiple LLM providers
litellm==1.12.1
deepseek==0.0.1
anthropic==0.8.1
